chore: rename date-type module to datetype
